requirements.txt
Django>=3.2.9,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2-binary>=2.9.1